# spec/oas_v1.yaml
openapi: 3.1.0
info:
  title: Agora's OP API (Beta)
  description: |-
    Welcome to the Collective API documentation. This API provides various endpoints to interact with the Collective platform, focusing on RetroPGF (Retroactive Public Goods Funding) for Optimism. Below is a comprehensive guide to the API endpoints, categorized by functionality.

    ### Current Status
    - **Beta**: This API is currently in beta. Endpoints marked as "Not Live" are under development and available for feedback only.

    ### API Categories
    Live and stable. 
    - **auth**: Authenticate with the Agora API
    - **Round 4**: Data related to Retro Funding for Optimism Round 4
    - **RetroFunding**: Data related to Retro Funding for Optimism Round 4 and beyond
    - **projects**: Data related to RetroPGF Projects
    - **RetroFundingRounds**: Data related to RetroPGF Funding Rounds
    - **RetroFundingBallots**: Data related to RetroPGF Funding Ballots
    - **impactMetrics**: Data related to the Impact Metrics for RetroFunding Round 4
    - **RetroFundingImpactMetrics**: Connect RetroPGF to Impact Metrics
    - **comments**: Data related to RetroPGF Project Comments
    - **commentsVotes**: Data related to RetroPGF Project Comments' Votes
    - **delegates**: Delegates data
    - **delegations**: Delegations data
    - **proposals**: Proposal data
    - **votes**: Vote data
    - **contracts**: Data for the current onchain contracts
    - **projects** Round 5 Projects with mock data
...

...

...

...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthToken:
      summary: Token for authentication
      description: >
        A token for authentication and authorization on the Agora platform.
        Subsequent requests should supply this as a Bearer token in the 
        "Authorization" header.
      type: object
      properties:
        access_token:
          summary: The access token
          description: The access token.
          type: string
        token_type:
          summary: Type of the token
          description: >
            The type of the token, either "jwt" or "api_key".
          type: string
        expires_in:
          summary: Expiration time of the token
          description: The time after which the token will expire, in seconds.
          type: integer
    SIWEVerificationBody:
      summary: Body of SIWE /verify request
      description: >
        Body SIWE verification requests containing message and signature per 
        (EIP-4361)[https://eips.ethereum.org/EIPS/eip-4361]
      type: object
      properties:
        message:
          summary: SIWE message for verification
          type: string
        signature:
          summary: SIWE message signature for verification
          type: string
        nonce:
          type: string
...
    RetroFundingBallot5ProjectsAllocation:
      summary: Projects allocation distribution based on the vote
      description: >
        Projects allocation distribution based on the vote out of total 10M OP
      type: object
      properties:
        project_id:
          type: string
        name:
          type: string
        image:
          type: string
        position:
          type: number
        allocation:
          type: number
        impact:
          type: number
          description: >
            The impact of the project on the ecosystem. This is a number between 0 and 5. 
            0 - Conflict of interest.
            1 - Very low impact.
            2 - Low impact.
            3 - Medium impact.
            4 - High impact.
            5 - Very high impact.
    RetroFundingBallotCategoriesAllocation:
      summary: Categories allocation distribution based on the vote
      description: >
        Categories allocation distribution based on the vote out of total 10M OP
      type: object
      properties:
        category_slug:
          type: string
          enum:
            - ETHEREUM_CORE_CONTRIBUTIONS
            - OP_STACK_RESEARCH_AND_DEVELOPMENT
            - OP_STACK_TOOLING
        allocation:
          type: string
        locked:
          type: boolean
...
    Round5Ballot:
      summary: A ballot for RetroFunding Round 5
      description: >
        A ballot for RetroFunding Round 5, including information about the ballot.
      type: object
      properties:
        address:
          summary: Address of the voter
          description: Address of the voter
          type: string
        round_id:
          type: integer
        status:
          type: string
          enum:
            - NOT STARTED
            - RANKED
            - PENDING SUBMISSION
            - SUBMITTED
        budget:
          description: >
            Budget allocation selected by voter. This is a integer between 2,000,000 and 8,000,000.
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
        category_allocations:
          type: array
          items:
            $ref: "#/components/schemas/RetroFundingBallotCategoriesAllocation"
        projects_allocations:
          type: array
          items:
            $ref: "#/components/schemas/RetroFundingBallot5ProjectsAllocation"
        projects_to_be_evaluated:
          description: >
            Projects that haven't been evaluated by the voter. This is a list of project IDs.
          type: array
          items:
            type: string
        total_projects:
          description: >
            Total number of projects that the voter has to evaluate. This is the length of the projects_to_be_evaluated array.
          type: integer
        payload_for_signature:
          $ref: "#/components/schemas/RetroFunding5BallotSubmissionContent"
    RetroFunding4BallotSubmission:
      summary: Fields required for ballot submission
      description: An object containing the information required to submit a ballot.
      type: object
      properties:
        address:
          summary: Address or ENS of the ballot caster
          description: Address or ENS of the ballot caster.
          type: string
        ballot_content:
          type: object
          properties:
            allocations:
              type: array
              items:
                type: object
                metricId:
                  type: number
            os_only:
              type: boolean
            os_multiplier:
              type: number
        signature:
          summary: Badgeholder signature of the ballot
          description: >
            Badgeholder's signature of the contents of their ballot.
          type: string
          pattern: "^0x[a-fA-F0-9]{130}$"
    RetroFunding5BallotSubmissionContent:
      summary: Fields required for ballot submission
      type: object
      properties:
        budget:
          type: number
        category_allocation:
          type: array
          items:
            type: object
            categoryId:
              type: number
        projects_allocation:
          type: array
          items:
            type: object
            projectId:
              type: number
    RetroFunding5BallotSubmission:
      summary: Fields required for ballot submission
      description: An object containing the information required to submit a ballot.
      type: object
      properties:
        address:
          summary: Address or ENS of the ballot caster
          description: Address or ENS of the ballot caster.
          type: string
        ballot_content:
          $ref: "#/components/schemas/RetroFunding5BallotSubmissionContent"
        signature:
          summary: Badgeholder signature of the ballot
          description: >
            Badgeholder's signature of the contents of their ballot.
          type: string
          pattern: "^0x[a-fA-F0-9]{130}$"
  parameters:
    addressOrEnsName:
      name: addressOrEnsName
      in: path
      description: The address or ENS name of the delegate to retrieve
      required: true
      schema:
        type: string
      examples:
        address:
          value: "0xDa6d1F091B672C0f9e215eB9fa6B5a84bF2c5e11"
          summary: Address of the delegate.
        ensName:
          value: "flipliquid.eth"
          summary: ENS name of the delegate.
...
    roundIdParam:
      name: roundId
      in: path
      description: The round ID of the RetroFunding round to retrieve.
      required: true
      schema:
        type: integer
...
    budgetParam:
      name: budget
      in: path
      description: The budget to be set in OP tokens
      required: true
      schema:
        type: number
        minimum: 2000000
        maximum: 8000000
paths:
  /spec:
    get:
      summary: Gets this specification
      description: Retrieves the full OAS/Swagger spec for the API in YAML.
      operationId: getSpec
      tags:
        - spec
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /auth/nonce:
    get:
      summary: Gets nonce for SIWE authentication
      description: Gets nonce for SIWE authentication.
      operationId: getNonce
      tags:
        - auth
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  /auth/verify:
    post:
      summary: Posts SIWE verification payload
      description: Posts SIWE message and signature.
      operationId: postSiweVerificationMessage
      tags:
        - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SIWEVerificationBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
...
  /retrofunding/rounds/{roundId}/ballots/{addressOrEnsName}/submit:
    post:
      summary: Submits a particular ballot
      description: Submits the content of a ballot to be counted as final for the round.
      operationId: submitRetroFundingBallot
      tags:
        - RetroFunding
        - RetroFundingBallots
        - Round 4
        - Round 5
      parameters:
        - $ref: "#/components/parameters/roundIdParam"
        - $ref: "#/components/parameters/addressOrEnsName"
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/RetroFunding4BallotSubmission"
                - $ref: "#/components/schemas/RetroFunding5BallotSubmission"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  round:
                    type: number
                  signature:
                    type: string
                  payload:
                    type: string
                  created_at:
                    type: string
                  updated_at:
                    type: string
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
...
  /retrofunding/rounds/{roundId}/ballots/{addressOrEnsName}/budget/{budget}:
    post:
      summary: Updates a budget allocation for a specific RetroFunding ballot
      description: >
        Updates a category budget allocation for a RetroFunding round on Agora. Value is in OP tokens. Min 2M, Max 8M.
      operationId: updateRetroFundingRoundBudgetAllocation
      tags:
        - RetroFunding
        - Round 5
      parameters:
        - $ref: "#/components/parameters/roundIdParam"
        - $ref: "#/components/parameters/addressOrEnsName"
        - $ref: "#/components/parameters/budgetParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round5Ballot"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
...